Finalized module system
Fixed some bugs.
Added many logging possibilities in order to debug properly.
Modules now get properly loaded and only enabled when they are installed
to the database and enabled by the user. A uninstalled/disabled module
will not enable. If there is any dependency that was not enabled, the
server won't start up at the moment.