New Session handling and Events
Removed the old WebSession classes and added the Authenticator interface
with the AuthenticatorFactory class. The AuthenticatorFactory provides
all Contexts with an Authenticator implementation that reads the data
from the Session or else. Implemented events using an EventManager and
Listener. Implemented the Servlet for managing settings. Many bug fixes
as well!