Working on MVC Support
Working on implementation of a servlet routing requests to a specific
method of a Controller instance annotated with several annotations.
The currently supported annotation provide a way to set the path
mapping, allowed request method, use of url and query parameters in
method signature as well as setting and checking the Accept Header.

Also added the AdminGroupListServlet and made some database changes.